refactor: centralize YouTube playlist helpers
